+++ src/components/Popover/index.js
@@ -1,0 +1,18 @@
+import Popover from './Popover';
+import { usePlugin, installPlugin } from './../../utils';
+
+const VuePlugin = {
+  install(Vue) {
+    installPlugin(Vue, Popover);
+
+    
+  },
+};
+
+usePlugin(VuePlugin);
+
+export default VuePlugin;
+
+export {
+  Popover,
+};