+++ src/docs/TestPopover.vue
@@ -1,0 +1,21 @@
+<template>
+  <z-card>
+    <z-card-header>
+      <z-card-title content="Google" :size="5" />
+      <z-card-subtitle content="Etiam tincidunt interdum risus" color="gray" />
+    </z-card-header>
+    <z-card-body>
+      <p>Nullam elementum fringilla faucibus. Maecenas massa augue, condimentum non neque eu, euismod accumsan turpis.</p>
+    </z-card-body>
+    <z-card-footer>
+      <z-button-group block>
+        <z-button type="primary">Button</z-button><z-button>Button</z-button>
+      </z-button-group>
+    </z-card-footer>
+  </z-card>
+</template>
+<script>
+export default {
+
+}
+</script>